[UPTADE][Linux]

+++ app/src/main/java/com/example/videomagnification/model/Params.java
@@ -2,13 +2,13 @@ package com.example.videomagnification.model;
 
 public class Params {
     private int alp0;
-    private float lamb0;
+    private int lamb0;
 
     public int getAlp0() {
         return alp0;
     }
 
-    public float getLamb0() {
+    public int getLamb0() {
         return lamb0;
     }
 }
